2.Basic_LLM_App_LCEL_FastAPI/requirements.txt
langchain
python-dotenv

langchain-openai
langchain-core

beautifulsoup4  # for web scraping
langchain_community

faiss-cpu  # for fast similarity search
streamlit  # for web app

langchain_core
langchain_ollama

#############################

langchain
langchain-groq
langchain_core
fastapi
uvicorn

langserve[all]